♻️ [Refactor] typeorm 폴더 분리
<footer>
- 관련: #3 #8
