+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { TypeOrmModuleOptions } from './configs';
+import { TypeOrmModuleOptions } from './configs/typeorm';
 import { BoardModule } from './modules/boards.module';
 
 @Module({
